Wraps up MongoDB Database Case Study
Case Study is escalated because of the following reasons:
* Insertion of data into DB does not work because
DBCollection.insert(DBObject ... arr) is not possible in current IL
implementation (Array of Complex Input Parameters on a
CallOperationAction).
* Retrieving the final query result does not work because
DBCursor.next() returns an interface. The current IL implementation
does not look for implementations of interfaces (interface return value
of a CallOperationAction is not supported).